move DB-related tests from new user tests to test_user_database.py
